requirements.txt
# HTTP and file handling
requests>=2.32.4
python-multipart>=0.0.6


# Utilities
python-dotenv>=1.1.1
typing-extensions

#Gemini Models
google-genai>=1.29.0
google-generativeai>=0.8.5
#For OSS Model
groq>=0.31.0

openai-harmony>=0.0.4